% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auto_palette}
\alias{auto_palette}
\title{Auto-generate color palette}
\usage{
auto_palette(n)
}
\arguments{
\item{n}{Number of colors to generate}
}
\value{
Character vector of hex color codes
}
\description{
Auto-generate color palette
}
\keyword{internal}
