% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{bezier_ribbon}
\alias{bezier_ribbon}
\title{Create bezier curve ribbon polygon}
\usage{
bezier_ribbon(sx0, sx1, sy, tx0, tx1, ty, curvature = 0.55, n = 250)
}
\arguments{
\item{sx0}{Source x start}

\item{sx1}{Source x end}

\item{sy}{Source y position}

\item{tx0}{Target x start}

\item{tx1}{Target x end}

\item{ty}{Target y position}

\item{curvature}{Curve strength (0-1), default 0.55}

\item{n}{Number of points for smoothness, default 250}
}
\value{
Data frame with x, y coordinates for polygon
}
\description{
Generates coordinates for a curved ribbon connecting two horizontal segments
using cubic Bezier curves.
}
\keyword{internal}
