% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Synteny_from_mcscanx.R
\name{build_blocks_from_anchors}
\alias{build_blocks_from_anchors}
\title{Build synteny blocks from anchor pairs}
\usage{
build_blocks_from_anchors(anchors)
}
\arguments{
\item{anchors}{A standardized anchor table.}
}
\value{
A standardized block table.
}
\description{
Build synteny blocks from anchor pairs
}
\examples{
NULL
}
