% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene.R
\name{compute_block_layout}
\alias{compute_block_layout}
\title{Compute linear block coordinates}
\usage{
compute_block_layout(block_df, chr_layout)
}
\arguments{
\item{block_df}{A block table produced by [build_blocks_from_anchors()].}

\item{chr_layout}{Output of [compute_chr_layout()].}
}
\value{
A tibble with linear coordinates for synteny blocks.
}
\description{
Compute linear block coordinates
}
\examples{
NULL
}
