% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene.R
\name{compute_chr_layout}
\alias{compute_chr_layout}
\title{Compute linear chromosome coordinates}
\usage{
compute_chr_layout(chr_df, gap_size = 1e+06)
}
\arguments{
\item{chr_df}{A data frame with `genome_id`, `chr_id`, and `length` columns.}

\item{gap_size}{Numeric gap inserted between chromosomes.}
}
\value{
A tibble with linear chromosome coordinates.
}
\description{
Compute linear chromosome coordinates
}
\examples{
NULL
}
