% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_analysis_design.R
\name{create_analysis_design}
\alias{create_analysis_design}
\title{Create a multi-genome analysis design}
\usage{
create_analysis_design(genomes, comparisons)
}
\arguments{
\item{genomes}{A genome registry produced by [register_genomes()].}

\item{comparisons}{A comparison registry produced by [define_comparisons()].}
}
\value{
A list with class `ggsynteny_design`.
}
\description{
Create a multi-genome analysis design
}
\examples{
NULL
}
