% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_representative_transcript.R
\name{get_representative_transcript}
\alias{get_representative_transcript}
\title{Extract representative protein transcript from a FASTA file}
\usage{
get_representative_transcript(fasta, sep, output)
}
\arguments{
\item{fasta}{Character
Path to the input amino acid FASTA file to be processed.}

\item{sep}{Character
Separator used in the sequence IDs to distinguish gene ID and transcript ID.}

\item{output}{Character
Path to the output FASTA file that will store the representative (longest) transcript per gene.}
}
\value{
None
}
\description{
Extract representative protein transcript from a FASTA file
}
\examples{
NULL

}
