% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggsynteny-package.R
\docType{package}
\name{ggsynteny-package}
\alias{ggsynteny}
\alias{ggsynteny-package}
\title{ggsynteny: Publication-Quality Synteny Plots}
\description{
Create publication-quality synteny visualizations for comparative genomics.
Supports both macro-synteny (chromosome-level) and micro-synteny (gene-level)
plots with flexible coloring schemes.
}
\section{Main Functions}{

\itemize{
\item \code{\link{plot_synteny}} - Create macro-synteny plots
\item \code{\link{plot_microsynteny}} - Create micro-synteny plots
}
}

\section{Colour Palettes}{

Every \code{palette} argument accepts, by name, all 32 palettes of the
\href{https://github.com/loukesio/ltc-color-palettes}{ltc package}
(vendored, so ltc need not be installed), e.g.
\code{plot_synteny(syn, sp_order, palette = "casa_natal")}.
List them with \code{\link{syn_palettes}}.
}

\section{Data Parsers}{

\itemize{
\item \code{\link{read_synteny_tsv}} - Read TSV format
\item \code{\link{read_mcscanx}} - Read MCScanX format
\item \code{\link{read_genespace}} - Read GENESPACE format
}
}

\section{Example Data}{

\itemize{
\item \code{\link{example_synteny_data}} - Macro-synteny example
\item \code{\link{example_microsynteny_data}} - Micro-synteny example
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/loukesio/ggsynteny}
  \item \url{https://loukesio.github.io/ggsynteny/}
  \item Report bugs at \url{https://github.com/loukesio/ggsynteny/issues}
}

}
\author{
\strong{Maintainer}: Loukas Theodosiou \email{loukesio@gmail.com} (\href{https://orcid.org/0000-0001-6418-4652}{ORCID})

}
\keyword{internal}
