% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_gene.R
\name{map_genes_to_linear}
\alias{map_genes_to_linear}
\title{Map genomic intervals to linear chromosome coordinates}
\usage{
map_genes_to_linear(genes_df, chr_layout)
}
\arguments{
\item{genes_df}{A data frame with `genome_id`, `chr_id`, `start`, and `end` columns.}

\item{chr_layout}{Output of [compute_chr_layout()].}
}
\value{
A tibble with linear coordinates.
}
\description{
Map genomic intervals to linear chromosome coordinates
}
\examples{
NULL
}
