% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gff.R
\name{parse_gff}
\alias{parse_gff}
\title{Extract chromosome lengths from a GFF3 annotation file}
\usage{
parse_gff(gffpath, output)
}
\arguments{
\item{gffpath}{Character
Path to the input GFF3 annotation file.}

\item{output}{Character
Path to the output file in MCScanX gff format.}
}
\value{
None
}
\description{
Extract chromosome lengths from a GFF3 annotation file
}
\examples{
NULL
}
