% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_gff3_to_mcscanx.R
\name{parse_gff3_to_mcscanx}
\alias{parse_gff3_to_mcscanx}
\title{Convert GFF3 annotation to MCScanX gff format}
\usage{
parse_gff3_to_mcscanx(gffpath, output)
}
\arguments{
\item{gffpath}{Character
Path to the input GFF3 annotation file.}

\item{output}{Character
Path to the output file in MCScanX gff format.}
}
\value{
None
}
\description{
Convert GFF3 annotation to MCScanX gff format
}
\examples{
NULL
}
