% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Synteny_from_mcscanx.R
\name{parse_Synteny_from_mcscanx}
\alias{parse_Synteny_from_mcscanx}
\title{Parse synteny blocks from MCScanX output}
\usage{
parse_Synteny_from_mcscanx(synteny, gff)
}
\arguments{
\item{synteny}{Character path to the MCScanX `.collinearity` file.}

\item{gff}{Character path to the MCScanX gff file.}
}
\value{
A legacy list containing block coordinates and chromosome lengths.
}
\description{
Parse synteny blocks from MCScanX output
}
\examples{
NULL
}
