% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Synteny_from_jcvi.R
\name{prepare_jcvi_data}
\alias{prepare_jcvi_data}
\title{Parse synteny blocks from jcvi anchors and BED files}
\usage{
prepare_jcvi_data(
  anchors,
  bed_a,
  bed_b,
  genome1 = "Genome1",
  genome2 = "Genome2",
  gap_size = 1e+06
)
}
\arguments{
\item{anchors}{Character path to a jcvi anchors file.}

\item{bed_a}{Character path to the first jcvi BED file.}

\item{bed_b}{Character path to the second jcvi BED file.}

\item{genome1}{Label for the upper genome in the plot.}

\item{genome2}{Label for the lower genome in the plot.}

\item{gap_size}{Numeric gap inserted between chromosomes when computing linear coordinates.}
}
\value{
A standardized `ggsynteny_data` object.
}
\description{
Parse synteny blocks from jcvi anchors and BED files
}
\examples{
NULL
}
