% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Synteny_from_mcscanx.R
\name{prepare_mcscanx_data}
\alias{prepare_mcscanx_data}
\title{Parse synteny blocks from an MCScanX collinearity file}
\usage{
prepare_mcscanx_data(
  synteny,
  gff,
  genome1 = "Genome1",
  genome2 = "Genome2",
  gap_size = 1e+06
)
}
\arguments{
\item{synteny}{Character path to the MCScanX `.collinearity` file.}

\item{gff}{Character path to the MCScanX gff file.}

\item{genome1}{Label for the upper genome in the plot.}

\item{genome2}{Label for the lower genome in the plot.}

\item{gap_size}{Numeric gap inserted between chromosomes when computing linear coordinates.}
}
\value{
A standardized `ggsynteny_data` object.
}
\description{
Parse synteny blocks from an MCScanX collinearity file
}
\examples{
NULL
}
