% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_analysis_inputs.R
\name{prepare_mcscanx_inputs}
\alias{prepare_mcscanx_inputs}
\title{Prepare ordered MCScanX inputs from a multi-genome analysis design}
\usage{
prepare_mcscanx_inputs(
  design,
  outdir,
  protein_sep = NULL,
  prepare_protein = TRUE
)
}
\arguments{
\item{design}{A design created by [create_analysis_design()].}

\item{outdir}{Character output directory for prepared files.}

\item{protein_sep}{Optional separator used to collapse transcript-level protein
FASTA IDs into gene-level representative sequences.}

\item{prepare_protein}{Logical, whether to generate representative protein FASTA
files when `protein_sep` is supplied.}
}
\value{
A list with class `ggsynteny_input_plan`.
}
\description{
Prepare ordered MCScanX inputs from a multi-genome analysis design
}
\examples{
NULL
}
