% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_parsers.R
\name{read_genespace}
\alias{read_genespace}
\title{Read GENESPACE/riparian output}
\usage{
read_genespace(synhits_file)
}
\arguments{
\item{synhits_file}{Path to synHits file}
}
\value{
List with elements \code{chromosomes} and \code{blocks}
}
\description{
Parses GENESPACE synHits file into ggsynteny format.
}
\details{
GENESPACE outputs a synHits file with columns: ofID1, ofID2, chr1, chr2,
start1, start2, end1, end2, genome1, genome2, etc.
}
\examples{
\dontrun{
syn <- read_genespace("synHits.tsv")
}

}
