% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Synteny_from_jcvi.R
\name{read_jcvi_bed}
\alias{read_jcvi_bed}
\title{Read a jcvi BED file}
\usage{
read_jcvi_bed(path, genome = NA_character_)
}
\arguments{
\item{path}{Character path to a jcvi BED file.}

\item{genome}{Optional genome label.}
}
\value{
A tibble with gene coordinates.
}
\description{
Read a jcvi BED file
}
\examples{
NULL
}
