% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_Synteny_from_mcscanx.R
\name{read_mcscanx_gff}
\alias{read_mcscanx_gff}
\title{Read an MCScanX gff file}
\usage{
read_mcscanx_gff(path, genome = NA_character_)
}
\arguments{
\item{path}{Character path to the MCScanX gff file.}

\item{genome}{Optional genome label.}
}
\value{
A tibble with gene coordinates.
}
\description{
Read an MCScanX gff file
}
\examples{
NULL
}
