% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_blastp_to_mcscanx.R
\name{run_blastp_to_mcscanx}
\alias{run_blastp_to_mcscanx}
\title{Run BLASTP for MCScanX synteny analysis}
\usage{
run_blastp_to_mcscanx(
  query,
  db,
  output,
  blastdir,
  outfmt = 6,
  threads = 6,
  num_alignments = 5,
  evalue = 1e-10
)
}
\arguments{
\item{query}{Character
Path to the protein FASTA file used as the BLASTP query.}

\item{db}{Character
Path to the protein FASTA file used to build the BLAST database.}

\item{output}{Character
Path to the output BLAST result file (typically `.blast`).}

\item{blastdir}{Character
Directory containing BLAST+ executables}

\item{outfmt}{Character
BLAST output format (default: 6, tabular format).}

\item{threads}{Numeric
Number of threads for BLASTP (default: 6).}

\item{num_alignments}{Numeric
Maximum number of alignments to report per query (default: 5).}

\item{evalue}{Numeric
E-value threshold for BLASTP search (default: 1e-10).}
}
\value{
a result of blastp
}
\description{
Run BLASTP for MCScanX synteny analysis
}
\examples{
NULL
}
