% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_MCScanX.R
\name{run_MCScanX}
\alias{run_MCScanX}
\title{Run MCScanX synteny analysis}
\usage{
run_MCScanX(prefix, MCScanXdir)
}
\arguments{
\item{prefix}{Character
The prefix of input file for MCScanX software}

\item{MCScanXdir}{Character
Path of MCScanX software}
}
\value{
None
}
\description{
Run MCScanX synteny analysis
}
\examples{
NULL

}
