% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/register_analysis_design.R
\name{validate_analysis_design}
\alias{validate_analysis_design}
\title{Validate a multi-genome analysis design}
\usage{
validate_analysis_design(design)
}
\arguments{
\item{design}{A design created by [create_analysis_design()].}
}
\value{
The validated design.
}
\description{
Validate a multi-genome analysis design
}
\examples{
NULL
}
