% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_ggsynteny_data.R
\name{validate_ggsynteny_data}
\alias{validate_ggsynteny_data}
\title{Validate a ggsynteny data object}
\usage{
validate_ggsynteny_data(data)
}
\arguments{
\item{data}{A `ggsynteny_data` object.}
}
\value{
The validated input object.
}
\description{
Validate a ggsynteny data object
}
\examples{
NULL
}
